requirements.txt
flask==2.0.1
numpy==1.21.2
scipy==1.7.1
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
matplotlib==3.4.3
opencv-python==4.5.3.56
tqdm==4.62.2
gunicorn==20.1.0
